Changed all in boxValues.h to be doubles for sending to PC

+++ boxValues.h
@@ -5,7 +5,7 @@ namespace boxValues{
 
 	struct returnData {
 		double v1;
-		short r1, r2;
+		double r1, r2;
 		double c1;
 		double f;
 	};
